Fix code style
